+++ imago/src/prefilter.h
@@ -36,6 +36,6 @@ namespace imago
    void _copyMatToImage (Image &img, const cv::Mat &mat);
 
    bool isCircle (Image &seg);
-   int estimateLineThickness(Image &bwimg, int grid = 10);
+   double estimateLineThickness(Image &bwimg, int grid = 10);
 }
 #endif /* _prefilter_h_ */
